gs_title: (s16) outer cast instead of (s32) for twin function stores
Removes extra extsh r0, r0 between lwz and sth when CW fctiwz result
is directly stored into *(s16*). (s16) cast on f32 expr naturally
truncates at sth. (s32) cast was adding redundant extsh.

- fn_80024F2C 92.8 -> 94.9
- fn_80024DBC 94.5 -> 96.6
- fn_8002509C 90.3 -> 92.4
- fn_8002520C 89.7 -> 90.5

+++ src/game/gs_title.c
@@ -758,8 +758,8 @@ LAB_80024E94:
         }
         break;
     }
-    *(s16*)(arg1 + 0x50) = (s32)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r28 + lbl_8047A374);
-    *(s16*)(arg1 + 0x52) = (s32)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r31 + lbl_8047A374);
+    *(s16*)(arg1 + 0x50) = (s16)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r28 + lbl_8047A374);
+    *(s16*)(arg1 + 0x52) = (s16)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r31 + lbl_8047A374);
 }
 #endif
 
@@ -828,8 +828,8 @@ LAB_80025004:
         }
         break;
     }
-    *(s16*)(arg1 + 0x50) = (s32)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r28 + lbl_8047A374);
-    *(s16*)(arg1 + 0x52) = (s32)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r31 - lbl_8047A374);
+    *(s16*)(arg1 + 0x50) = (s16)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r28 + lbl_8047A374);
+    *(s16*)(arg1 + 0x52) = (s16)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r31 - lbl_8047A374);
 }
 #endif
 
@@ -898,8 +898,8 @@ LAB_80025174:
         }
         break;
     }
-    *(s16*)(arg1 + 0x50) = (s32)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r28 - lbl_8047A374);
-    *(s16*)(arg1 + 0x52) = (s32)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r31 + lbl_8047A374);
+    *(s16*)(arg1 + 0x50) = (s16)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r28 - lbl_8047A374);
+    *(s16*)(arg1 + 0x52) = (s16)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r31 + lbl_8047A374);
 }
 #endif
 
@@ -968,8 +968,8 @@ LAB_800252E4:
         }
         break;
     }
-    *(s16*)(arg1 + 0x50) = (s32)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r28 - lbl_8047A374);
-    *(s16*)(arg1 + 0x52) = (s32)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r31 - lbl_8047A374);
+    *(s16*)(arg1 + 0x50) = (s16)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r28 - lbl_8047A374);
+    *(s16*)(arg1 + 0x52) = (s16)(lbl_8047B8E0 + (f32)(s32)(s16)unaff_r31 - lbl_8047A374);
 }
 #endif
